identity-management-service-web-api/src/main/java/au/id/hazelwood/idms/web/filter/CrossOriginRequestFilter.java
/*
 * Licensed to the Apache Software Foundation (ASF) under one or more
 * contributor license agreements.  See the NOTICE file distributed with
 * this work for additional information regarding copyright ownership.
 * The ASF licenses this file to You under the Apache License, Version 2.0
 * (the "License"); you may not use this file except in compliance with
 * the License.  You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package au.id.hazelwood.idms.web.filter;

import org.apache.commons.lang3.StringUtils;
import org.springframework.web.filter.OncePerRequestFilter;

import javax.servlet.FilterChain;
import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

/**
 * Filter to add cross origin request headers on all requests.
 *
 * @author Ricky Hazelwood
 */
public class CrossOriginRequestFilter extends OncePerRequestFilter
{
    private String allowedMethods = "GET, HEAD, POST, PUT, DELETE, TRACE, OPTIONS, PATCH";

    @Override
    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
        throws ServletException, IOException
    {
        String origin = request.getHeader("Origin");
        if (StringUtils.isNotEmpty(origin))
        {
            response.setHeader("Access-Control-Allow-Origin", origin);
            response.setHeader("Access-Control-Allow-Headers", request.getHeader("Access-Control-Request-Headers"));
            response.setHeader("Access-Control-Allow-Methods", this.allowedMethods);
            response.setHeader("Access-Control-Max-Age", "3600");
        }
        filterChain.doFilter(request, response);
    }

    public void setAllowedMethods(String allowedMethods)
    {
        this.allowedMethods = allowedMethods;
    }
}
